refactor(backend): replace float with Decimal across all financial computations
Systematic conversion of all float() usage to Decimal for financial values
per CLAUDE.md mandate ("Decimal always — never float"). Covers 29 backend
files (engines, services, repositories, schemas, API layer) and 3 frontend
files (mobile overflow fixes from QA report).

Key changes:
- Engines (QFS, FSAS, matrix, tier): all thresholds/weights as Decimal constants
- Services: Decimal arithmetic with ROUND_HALF_UP quantization
- Repositories: preserve DB Decimal values, no float() conversion
- Schemas: Pydantic models use Decimal fields for all financial data
- API layer: _safe_float → _safe_decimal
- Frontend: fix mobile horizontal overflow on all pages

+++ backend/app/models/schemas/signals.py
@@ -6,6 +6,7 @@ v2: Added bulk update schemas and signal change history.
 """
 
 from datetime import date, datetime
+from decimal import Decimal
 from typing import Optional
 from uuid import UUID
 
@@ -51,7 +52,7 @@ class SignalResponse(BaseModel):
     sector_name: str
     signal: str
     confidence: str
-    signal_weight: float
+    signal_weight: Decimal
     effective_date: date
     updated_by: str
     notes: Optional[str] = None
@@ -117,7 +118,7 @@ class SectorWithSignal(BaseModel):
     sector_name: str
     signal: Optional[str] = None
     confidence: Optional[str] = None
-    signal_weight: Optional[float] = None
+    signal_weight: Optional[Decimal] = None
     notes: Optional[str] = None
     effective_date: Optional[date] = None
     updated_by: Optional[str] = None
